[RUMF-83] fix View creation when both logs and RUM bundles are present
When both bundles were present, the logs bundle was initializing the
session, so the RUM bundle wasn't aware and didn't notify the
renewSession lifeCycle event.

This commit factorize the session handling and makes sure the "renew
session" events is correctly notified when the sessionId changes.

It also add an E2E test for this particular use-case.