[RUMF-201] add debug monitoring for navigation timing entries (#214)
* 🔥 [RUMF-201] remove debug log

* 🔊 [RUMF-201] add debug log to log the PerformanceNavigationTiming

This information will help us debug the issue by providing more context
on the navigation entry (ex: its name, initiatorType, ...) and will show
previous measures if any.  We expect only one
PerformanceNavigationTiming should happen when a page is loading, but
maybe that's not the case, so previous measures will tell us if it's
happening.