:bug:[RUMF-293][fetch] handle fetch response text error
When consuming a fetch response body, it can be disturbed or locked, a rejected promise is then returned (cf https://fetch.spec.whatwg.org/#concept-body-consume-body).
This rejected promise was displayed on the browser console and the corresponding request was not collected.

Handle the rejected promise and add the corresponding error to the collected request response.