Automatic click user action collection (#338)
* ✨ [RUM] collect request start

* ✨ [RUM] collect DOM mutations

* ✅ improve e2e types

This commit addresses typings issues in the E2E tests to make tests
more easier to write.

RUM events returned by the server are not the same as those used accross
the sdk. Typically, their properties are snake cased instead of
camelCased.  To improve this, we remove the SDK types usages, and
replace them by custom types used only for those tests.  A bit of
typings are repeated, but it makes the DX more easier.  Those types can
be adjusted at will for the tests needs.

Some global types have been adjusted too.