[RUMF-441] Track event counts for user actions (#358)
* 🏷  [RUM] moveUserAction related types to userActionCollection

* [RUM] rename 'newUserAction' to 'waitUserActionCompletion'

This new name is more explicit on what the function really those. In the
next commit, the function will be trimed down a bit to focus on the
'wait' process.

* [RUM] extract the user action plumbing into a function

This new `newUserAction` function uses other, self-contained functions
to implement a single user action collection.  The goal is to reuse this
for the future 'view load' user actions.

* 🎨 [RUM] move userActionCollection functions to have a consistent order

First public functions, then private function in the same order they
appear in the public functions.

This commit also moves the associated tests to follow the same order.

* [RUM] externalize trackEventCounts

* ✨ [RUM] track event counts occuring during a user action

* 👌✅ [RUM] use time constants in userActionCollection spec

* 👌 [RUM] fix events count reset between views

This fix is not very elegant.  I did work on a nicer fix, but it
involved changing the viewTracker implementation.  I may open a separate
PR for this after a bit more work.

* 👌✅ [RUM] use eventsCollector in more places

* 👌✅ [RUM] remove object literal type assertions

This is a workaround on the no-object-literal-type-assertion.  This rule
is useful in general, but in tests it can be cumbersome to mock the full
objects every time.  We can't disable this rule just for tests for now,
but we'll be able to do so when switching to ESLint

* 👌✅ [RUM] factorize lifecycle creation

* 👌 [RUM] fix doc typo

* 👌✅ [RUM] rename getEventCount

* 👌 [RUM] rename trackMeasures, extract the event counts tracking logic