[RUM] remove session handling in view logic
This commit takes a different approach on fixing the session id used
when sending RUM events right before the session is renewed.

Originally, it was fixed in commit
7bc0dfbb99d147ac04e8be5cc129146da811b21a by storing the session id into
the current view.  Since the view was recreated when the session was
renewed, events sent during a view were guaranteed to have the same
session id than the view.

The problem with this approach was that the sessionId was unexpectedly
stored in the view context, making different parts of the code
dependendent to the view logic.

This new solution goal is to be more explicit, and simplify the view
logic. The current session id (and type) is kept into the rumSession
module, to ensure a consistency accross the RUM library. A new lifeCycle
event is sent when the view is about to be renewed, but the session id
is still the same.