✅ [RUMF-759] cleanup RUM unit tests specHelper (#605)
* [RUMF-759] use RUM_EVENT_COLLECTED events to test sent events

With the newly introduced RUM_EVENT_COLLECTED and
RUM_EVENT_V2_COLLECTED, we can test for generated events without the
need to retrieve them from a mocked server. This simplifies the unit
tests a bit.

* [RUMF-759] introduce withConfiguration

This new helper allows to edit the configuration before any
`beforeBuild` callback is invoked.  It slightly simplifies the FF
edition.

* [RUMF-759] replace the performance observer stub with a spy

* [RUMF-759] use an object for beforeBuild

This will allow us to provide more data for beforeBuild callbacks

* [RUMF-759] allow to return a cleanup callback

A few spec helper methods that could be inlined in their spec are using
a cleanup function. To allow moving them, we add a way to define a
cleanup function in beforeBuild callbacks.

* [RUMF-759] inline withViewCollection in trackViews spec

* [RUMF-759] inline withActionCollection in the trackActions spec

* [RUMF-759] inline withPerformanceCollection in the performanceCollection spec

* [RUMF-759] inline withParentContexts() in parentContexts spec

* [RUMF-759] inline withRum in rum spec

* [RUMF-759] inline withInternalContext in internalContext spec

* [RUMF-759] inline withAssembly / withAssemblyV2 in assembly specs

* restore oneFailurePerSpec: true

With this value to false, the tests are kind of failing on IE where we
mark some test suite as "pending" in the `beforeEach` callback.  This is
a known issue in Jasmine https://github.com/jasmine/jasmine/issues/1579

* 👌 return an object with a `stop` method for cleanup