⬆️ update sinon (#672)
* ⬆️ update sinon

Sinon changed its build process and now provide the same code for 'main'
and 'modules' entry point, only the export syntax changes (as it should
be!).

We can remove the webpack hack for browserstack unit tests.

This allows running unit tests on IE with the 'local' configuration.

* ✅ don't automatically include any type for test app

When checking for TS compatibility, we need all dev dependencies to
build various libraries.  @types/sinon gets automatically included
somehow, and it is not comtible with TS 3.0 anymore.