👷 use paths from tsconfig.base.ts on build:bundle
The rum-recorder build:bundle was broken because it didn't define the
'@datadog/browser-rum-core' path (since it is now an indirect
dependency).

Instead of adding it, this commit removes all paths contained in
tsconfig packages, and configure the tsconfig-paths-webpack-plugin to
read paths from the `tsconfig.base.json` file, to avoid too many
duplicated paths configuration.