👷 simplify bundle configuration
This commit simplifies the bundle process a bit further:

* Each bundled package is now in charge of generating suffixed files on
its own.

* The WITH_SUFFIX environment variable is removed.

* The replace-build-env script is no more applied to bundled packages.
Instead, string-replace-loader is used to replace the 'build env', and
BannerPlugin is used to automatically add a banner to suffixed bundles.
It is still used for TypeScript generated files though.

* The deploy process is now a bit simpler too, since all bundles are
generated at once, everything can be deployed at the same time.