recorder: introduce helpers to obfuscate elements
Adds the nodeIsHidden & nodeOrAncestorsAreHidden helpers.

These helpers, part of the new "privacy" file, replace rrweb's
isBlocked() and rrweb-snapshot's _isBlockedElement() functions.

They do not rely on the blockClass variable (removed in a later
commit), but instead use constants defined within the same file.

An element is considered hidden if it has the data attribute
"data-dd-privacy" set to "hidden" or a class of "dd-privacy-hidden".