recorder: introduce helpers to ignore input
Adds the nodeHasInputIngored & nodeOrAncestorsHaveInputIngnored
helpers.

These helpers will return true when a node should have it's input
ignored, either because an attribute/class is set, or because it's a
kind of input we don't want to track.

They do not rely on the ignoreClass variable (removed in a later
commit), but instead use constants defined within the same file.

An element is considered ignored if it has the data attribute
"data-dd-privacy" set to "input-ignored", a class of
"dd-privacy-input-ingored", or a type of "email", "password" or "tel".