🏗 remove global MutationBuffer instance
This commit separates the API used by `record()` from the
`MutationBuffer` class.  This allows `record()` to instanciate a simple
`MutationController` and pass it to the mutation observer logic to
control it.

Since the `MutationBuffer` instance is now created when starting to
observe mutations, we don't need to `init()` it after its creation,
making it safer to use.