⬆️ upgrade webpack and related
webpack 4 requires an old version of terser-webpack-plugin, which
requires ssri 6. To upgrade ssri to 8.0.1, we need to upgrade webpack to
5.

This forces us to upgrade a bunch of webpack-related dependencies to
support the new major version. A bit of configuration changes was also
needed to support the new version, most importantly for tests
(karma-webpack).