🐛 [RUMF-882] fix the max_size flush issue
This commit changes the logic around DeflateSegmentWriter. Instead of
having a single, global DeflateSegmentWriter for all segments, a new
DeflateSegmentWriter is created for each segment, handling all
communications with the DeflateWorker for this specific segment only.
This allows to use the Segment instance in onWrote and onFlushed
callbacks, and simplifies the asynchronous logic:

* the 'meta' object is not handled by the DeflateSegmentWriter anymore,
since we can just ask the Segment for its meta in the onFlushed callback

* and last but not least, we can check if the segment has already been
flushed before creating a new one if the size has reached the limit in
the onWrote callback.