✅ [RUMF-882] adjust the MockWorker to have properly encoded segments
Previously, the MockWorker didn't actually use the message data sent to
him, and responded with a Uint8Array of 0 bytes.

Since we will want to test what is actually built by the Segment, this
commit adjust the MockWorker implementation to be a bit closer to the
reality and respond with a Uint8Array that represent the encoded segment
(as UTF-8 instead of Deflate for simplicity).

Because the responses are now based on actual message data, we can't
mock the size like previously. Instead, we add a big record to make
segments reach their size limit, which is closer to the reality and
easier to understand.