🐛 [RUMF-882] Segments can be flushed because of the max_size limit even if the max_size isn't reached  (#787)
* ✅ adjust the mock DeflateWorker

To allow testing new scenarios:

* Allow more than one listener
* Add a way to process only one message at a time
* Add a way to skip one message (avoid using "id" as the concept of id
will change)

* [RUMF-882] add a test to illustrate the issue

* [RUMF-882] move monitoring of DeflateWorker outside of DeflateSegmentWriter

To fix the flush issue, we'll use a separate DeflateSegmentWriter for
each Segment instance. Multiple DeflateSegmentWriter could then be
listening for worker messages. We don't want to report errors multiple
times in this case, so move the monitoring directly on the DeflateWorker
singleton.

* 🐛 [RUMF-882] fix the max_size flush issue

This commit changes the logic around DeflateSegmentWriter. Instead of
having a single, global DeflateSegmentWriter for all segments, a new
DeflateSegmentWriter is created for each segment, handling all
communications with the DeflateWorker for this specific segment only.
This allows to use the Segment instance in onWrote and onFlushed
callbacks, and simplifies the asynchronous logic:

* the 'meta' object is not handled by the DeflateSegmentWriter anymore,
since we can just ask the Segment for its meta in the onFlushed callback

* and last but not least, we can check if the segment has already been
flushed before creating a new one if the size has reached the limit in
the onWrote callback.

* 🎨👌 [RUMF-882] adjust DeflateSegmentWriter listener structure

* ✅ [RUMF-882] adjust the MockWorker to have properly encoded segments

Previously, the MockWorker didn't actually use the message data sent to
him, and responded with a Uint8Array of 0 bytes.

Since we will want to test what is actually built by the Segment, this
commit adjust the MockWorker implementation to be a bit closer to the
reality and respond with a Uint8Array that represent the encoded segment
(as UTF-8 instead of Deflate for simplicity).

Because the responses are now based on actual message data, we can't
mock the size like previously. Instead, we add a big record to make
segments reach their size limit, which is closer to the reality and
easier to understand.

* ✅👌 [RUMF-882] rename MockWorker processAll/processOne/skipOne

* 👌🏗 [RUMF-882] merge DeflateSegmentWriter and Segment

* ✅ [RUMF-882] replace segmentFlushSpy with sendSpy check

To avoid confusing spy that are invoked at flush request
(segmentFlushSpy) and flush response (sendSpy), this commit replace the
usage of segmentFlushSpy with sendSpy. To do this, we need to make the
worker process a bit more messages than previously.

* ✅ [RUMF-882] clarify non-regression test

* Add a bit of comments to justify what the test is doing
* Introduce two other tests to illustrate the expected behavior of the
logic related to the issue.

* ✅ fix IE11 support

Build the TextEncoder when it is actually used.