[RUMF-895] cleanup the global "mirror" variable (#801)
* 🚚 [RUMF-895] introduce snapshot utils 'hasSerializedNode'

* replaces the previous `isINode` function
* use it in more places

* 🚚 [RUMF-895] introduce snapshot utils 'getSerializedNodeId'

* replaces mirror.getId
* accepts any "Node", so typings can be simplified a bit

* 🔥 [RUMF-895] remove unused function mirror.getNode

* [RUMF-895] introduce snapshot utils 'setSerialisedNode'

To collocate "serialized nodes" storage, create a new setSerialisedNode
in the same file as other functions that access node.__sn. This will
makes things a bit easier to follow.

* 🏗 [RUMF-895] introduce recursivelyRemoveSerializedNodes

Instead of using `mirror.map` to know if a DOM Node was injected in the
document or not, we will base simply remove the serialised nodes
attached to removed DOM node. This should be almost functionaly the same
as before, the only thing that might change is that we won't re-use the
id of a DOM node that has previously been injected then removed (in a
separate mutation batched)

* 🔥 [RUMF-895] remove now unused mirror/map

* 🏷️ [RUMF-895] rename 'INode' and make it more private

All serialized node accesses are done by the snapshot utils, there is no
need to use this type elsewhere anymore.

* ✅ [RUMF-895] add a few snapshot utils tests

* 👌 remove 'addNodeToMap'

* 👌 clarify and fix why recursivelyRemoveSerializedNodes is needed

* 👌 introduce a helper to avoid testing node ids against '-1'

* 👌 fix comment typo

* 👌 remove 'no-underscore-dangle' ESLint rule

* 👌🚚 rename utils to serializationUtils