🐛 fix removed node check
Before #801, `mirror.map` was used to track Nodes that have been removed
in a mutation (whether they were re-added in a subsequent mutation of
the same batch or not), to skip attributes and text mutations occuring
on those nodes.

With #801, we decided to remove `mirror.map`, and instead remove the
serialized node of removed nodes at the end of a mutation batch.

But, the current mutation processing logic expects to keep attached
serialized nodes between mutation batches. So we should not remove
serialized nodes.

Instead, let's just check if the node is still in the document before
taking the mutation into account.