🔥 remove mutation freeze logic (#802)
This logic was introduced to allow RRWeb users to "pause" the recording
by using the public method `freezePage`. This method has been removed in
PR #727, and it appears that the `freeze` logic is now unneeded: since
the "fullsnapshot" logic is synchronous, no DOM mutation can happen or
be processed while the document snapshot is taken.