🎨✅ [RUMF-902] make a more explicit API for validating mutations
The new API uses explicit functions to select and create new nodes. This
makes tests more readable, especially when there is multiple new nodes.
It also allows to make assertions on nested serialized nodes, and use it
without a full Segment or Records (so it can be used at deeper levels in
Unit tests)