🏗 [RUMF-902] introduce serializationUtils
This commit introduces a few utility functions to store and retrieve
serialized node ids stored in DOM nodes. The equivalent tools (`mirror`,
`INode`...) that are only used by the 'mutation' module are intentionally
left as-is, and will be removed in a future PR when we'll enable the new
mutation observer.

Before this commit, some 'observer' strategies didn't check if
`mirror.getId` returned a valid `id` before emitting a record. In
practice it shouldn't happen, but in unit tests it did happen because we
artificially emitted 'input' events on a DOM node that wasn't in the
document. This is why some tests were slightly adjusted to reflect the
reality more accurately.