✅ [RUMF-902] prepare tests for mutations collections rework (#808)
* 🚚 [RUMF-902] move recorder helpers from E2E to Unit

Because the unit tests have no concept of "EventRegistry", we pass the
segment directly instead of passing the registry.

* 🎨✅ [RUMF-902] make a more explicit API for validating mutations

The new API uses explicit functions to select and create new nodes. This
makes tests more readable, especially when there is multiple new nodes.
It also allows to make assertions on nested serialized nodes, and use it
without a full Segment or Records (so it can be used at deeper levels in
Unit tests)

* ✅ [RUMF-902] use the mutationPayloadValidator in record spec

This commit adjusts imported utilities from E2E to be used without a
Segment or Record, so it can be used in the record spec.

* ✅ [RUMF-902] simplify mutation spec

Mutation specs will become more important when we'll write more tests
for it. This commit simplifies the existing tests:

* it removes the MutationObserver mock. Having to mock mutations
manually was error prone, because we had to make the actual DOM changes,
then carefully craft mutation records corresponding to the changes we
made. By using actual MutationObservers, we only need to make DOM
changes, and the records are generated by the browser for free.

* it uses the "mutationPayloadValidator" to simplify the mutation record
validation

* it serializes the whole document first, like the actual recording
process would do during full snapshots

* apply comments suggestions

Co-authored-by: Bastien Caudan <bastien.caudan@datadoghq.com>

* 👌 adjust comments of changed functions

* 👌✅ add a helper to get the first segment from E2E events

* 👌✅ rename newNode/selectNode

Co-authored-by: Bastien Caudan <bastien.caudan@datadoghq.com>