👌 remove unneeded record.spec unit test
We chose to remove this test because its intent is a bit obscure and the
new mutation observer has a good enough test coverage. I kept the test
for the old mutation observer implementation since its coverage is still
lacking. It will be removed when we enable the new mutation observer.