⚗ [RUMF-902] implement a new mutation observer  (#810)
* 🏗 [RUMF-902] introduce serializationUtils

This commit introduces a few utility functions to store and retrieve
serialized node ids stored in DOM nodes. The equivalent tools (`mirror`,
`INode`...) that are only used by the 'mutation' module are intentionally
left as-is, and will be removed in a future PR when we'll enable the new
mutation observer.

Before this commit, some 'observer' strategies didn't check if
`mirror.getId` returned a valid `id` before emitting a record. In
practice it shouldn't happen, but in unit tests it did happen because we
artificially emitted 'input' events on a DOM node that wasn't in the
document. This is why some tests were slightly adjusted to reflect the
reality more accurately.

* ⚗ [RUMF-902] implement a new mutation observer

* 🚩 [RUMF-902] use the new mutation observer via a FF

* 👌 apply some suggestions from code review

Co-authored-by: Bastien Caudan <bastien.caudan@datadoghq.com>

* rename utils.spec to serializationUtils.spec

* 👌 remove undefined properties when using existing nodes more thoroughly

* 👌 remove the "handleMutations" function

* 👌 make sure 'nodeIsIgnored' is only called with serialized nodes

* 👌 rename a few things in the mutationObserver.spec

* 👌 clarify some mutationObserver tests

* 👌 remove useless uses of [0] for unused snapshot results

* 👌 move and rename sortNodesByTopologicalOrder

* 👌 remove unneeded record.spec unit test

We chose to remove this test because its intent is a bit obscure and the
new mutation observer has a good enough test coverage. I kept the test
for the old mutation observer implementation since its coverage is still
lacking. It will be removed when we enable the new mutation observer.

* Update record spec test title

Co-authored-by: Bastien Caudan <bastien.caudan@datadoghq.com>

Co-authored-by: Bastien Caudan <bastien.caudan@datadoghq.com>