✨ [RUMF-901] allow to run the 'performances' script with rum-recorder  (#818)
* 🐛 [RUMF-901] improve network profiling

The devtools protocol does not report request body if they are above a
few kb. To work around this limitation, this commit implements a HTTPS
server used as a Browser SDK "proxy", capturing the size of all requests
sent by the SDK.

* ✨ [RUMF-901] allow to run the 'performances' script with rum-recorder

* 👌 add the details of what was downloaded by the app

The idea is to be able to compare approximately the downloaded resources
size that may have an impact on the session-replay size (HTML,
stylesheet and scripts (kind of)) to the data that have actually been
uploaded to the session-replay intake.

* Revert "👌 add the details of what was downloaded by the app"

This reverts commit 6d8657b17520ea48f82af56f17ea493aefad8fdc.

Results are hard to interpret and adds a bit of noise, let's just remove
it for now.