✅ make the stylesheet record spec independent from DOM mutations
In this particular test, we want to ensure that stylesheet APIs are
correctly observed. Because it uses a setTimeout, processing mutations
asynchronously broke the test because the appended stylesheet was taken
into account after APIs were used. By moving the stylesheet creation
before starting to record, we avoid relying on the mutation processing.