✅ fix flaky test
The `setTimeout` made the test flaky because sometimes mutations were
processed before the 'span' node was added.

We don't really need it to ensure mutations are processed
asynchronously: we already ensure that the mutation callback is not
called synchronously right after adding an element. Batching mutations
that are triggered a bit later was interesting, but it is hard to test
without being flaky. Other tests are here to ensure this should not
cause any problem though.