⚡️[RUMF-896] process mutations asynchronously  (#832)
* 🚨 allow TS module declaration and Logical Expressions in module bodies

* ✅ make the stylesheet record spec independent from DOM mutations

In this particular test, we want to ensure that stylesheet APIs are
correctly observed. Because it uses a setTimeout, processing mutations
asynchronously broke the test because the appended stylesheet was taken
into account after APIs were used. By moving the stylesheet creation
before starting to record, we avoid relying on the mutation processing.

* ⚡️[RUMF-896] process mutations asynchronously

* 🐛 fix SSR compatibility

Move the 'window' accesses into a function.

* ✅ fix flaky test

The `setTimeout` made the test flaky because sometimes mutations were
processed before the 'span' node was added.

We don't really need it to ensure mutations are processed
asynchronously: we already ensure that the mutation callback is not
called synchronously right after adding an element. Batching mutations
that are triggered a bit later was interesting, but it is hard to test
without being flaky. Other tests are here to ensure this should not
cause any problem though.

* 👌 fix variable names

* 👌 move the mutations batching logic in its own module

* 👌 replace the pendingMutations array instead of emptying it