📦 [RUMF-905] update vulnerable dependencies (#836)
* 📦 [RUMF-905] update vulnerable dependencies

A few subdependencies have been updated to address new security
advisories. Notes:

* Bumbag uses fixed versions for its dependencies, and didn't release a
version with the new lodash version, so we hane to force the lodash
version there. This has been reported [here][1]

* `yarn audit` will still warn about a "moderate" vulnerability in the hosted-git-info
subdependency, but this is a false-positive: the author backported the
fix in version 2.8.9[2], but the security advisory hasn't been updated
with this information.

[1]: https://github.com/bumbag/bumbag-ui/issues/105#issuecomment-834398613
[2]: https://github.com/npm/hosted-git-info/blob/v2/CHANGELOG.md#289-2021-04-07

* 🏷️fix typecheck issue