♻️ cleanup recorder DOM serialization (#851)
* 🚚 move functions to make URLs absolute in `serializationUtils`

* ♻️ rename a few functions

* ♻️ cleanup transformAttribute functions a bit

* 🎨 reorder serialize module a bit

* ♻️ introduce common serialization options

* ♻️ split serializeNode function

* ♻️ serialize child nodes in `serializeNode`

This simplify the code structure, since we don't need to compute when we
should serialize child nodes outside of the parent serialization context.

* ♻️ do not serialize ignored node

Serializing ignored node is useless since we will never send the
serialized version. Instead, simply don't attribute an id to ignored
nodes, and adjust the condition to skip any node without id when
processing mutations.

* ♻️ replace "IdNodeMap" with a Set

A "Set" is better suited for the job, and a bit easier to understand.

* ✅ write serializeNodeWithId tests

* ✅ fix recorder unit tests failing with invalid expected id

Because the serialized version of 'document' and 'document type' nodes
were not cleaned up between tests, the id of new nodes was sometimes
incorrect because it used the uncleaned node ids as "maximum id" and
incremented from there.

* round scrollTop/Left so tests are passing on Chrome Android

* 👌 add a few tests for "makeUrlAbsolute"

* 🐛 fix ABSOLUTE_URL regex

The ABSOLUTE_URL didn't take protocols with multiple letters `http:`
into account. This was working nonetheless, because `makeUrlAbsolute`
allows to pass absolute URLs anyway.

* 👌 rename 'n' parameters to 'node'

* ✅ add a mutationObserver test about moving ignored nodes

* Update test title

Co-authored-by: Matthieu "Vlad" Hauglustaine <vlad@datadoghq.com>

Co-authored-by: Matthieu "Vlad" Hauglustaine <vlad@datadoghq.com>