✅ [RUMF-938] improve test
The test was inexact because we wanted to test the behavior when there
is no session, but `startRumSession` is creating a session if there is
none.

The test was working most of the time because `sampleRate` was 0.5%, so
calling `startRumSession` had a very low probability to generate a
tracked session. This commit changes the sample rate to 50% so tests
flakiness should be more obvious.