🔥 [RUMF-940] remove session.isTrackedWithResource support
As discussed, the "replaySampleRate" option will replace the
"resourceSampleRate" option since it will re-use the same value in the
session cookie.

As a first step to introduce the "replaySampleRate" option, let's remove
internal code related to resource sample rate.

This commit intentionally keeps the "resourceSampleRate" configuration
option. It will be replaced by the "replaySampleRate" in the next
commit.