[RUMF-998] implement a shouldMaskNode function
The goal is to simplify the Privacy Level typings by removing the
distinction between "Internal" and non-"Internal" levels. This is a
first step: it introduces a new function that focuses on handling the
"MASK_FORMS_ONLY" level. This way, we don't have to worry whether "MASK"
or "MASK_FORMS_ONLY" is transmitted to children.