[RUMF-998] getNodePrivacyLevel takes the initialPrivacyLevel as argument
This change is in preparation of the initialPrivacyLevel configuration
option. It changes the `getNodePrivacyLevel` paradigm a bit: instead of
receiving the direct parent privacy level, it receives a default privacy
level to be used if no level was found.

Thus, this function now always recurse over the node ancestors. During
serialization, where we know the parent privacy level, we can use the
`getNodeSelfPrivacyLevel` function to avoid this behavior.