✨ [RUMF-998] introduce the initialPrivacyLevel configuration option  (#1004)
* [RUMF-998] implement a shouldMaskNode function

The goal is to simplify the Privacy Level typings by removing the
distinction between "Internal" and non-"Internal" levels. This is a
first step: it introduces a new function that focuses on handling the
"MASK_FORMS_ONLY" level. This way, we don't have to worry whether "MASK"
or "MASK_FORMS_ONLY" is transmitted to children.

* [RUMF-998] remove remap function

This commit removes the "remap" function and associated. The
`getInternalNodePrivacyLevel` becomes `getNodePrivacyLevel` and can
return any privacy level.

This makes typings a bit less safe, but following commits will improve
the situation.

* [RUMF-998] rename NodePrivacyLevelInternal to NodePrivacyLevel

Now that associated functions have been removed, we can have a single
type for all privacy levels.

* [RUMF-998] getNodePrivacyLevel takes the initialPrivacyLevel as argument

This change is in preparation of the initialPrivacyLevel configuration
option. It changes the `getNodePrivacyLevel` paradigm a bit: instead of
receiving the direct parent privacy level, it receives a default privacy
level to be used if no level was found.

Thus, this function now always recurse over the node ancestors. During
serialization, where we know the parent privacy level, we can use the
`getNodeSelfPrivacyLevel` function to avoid this behavior.

* [RUMF-998] remove the "NOT_SET" privacy level variant

This reduce corner cases a bit by leveraging the type to restrict some
API usages.

* [RUMF-998] ensure only a subset of privacy levels can be inherited

Same as previous commit, this allows to reduce the scope of some
functions by passing explicit subsets of privacy levels

* [RUMF-998] make initial privacy level arguments required

To make sure the initial privacy level is correctly set everywhere, make
it required on `getNodePrivacyLevel` and `serializeDocument` (and
associated subfunctions).

* [RUMF-998] rename NodePrivacyLevel values to match public configuration

* ✨ [RUMF-998] introduce the initialPrivacyLevel configuration option

* [RUMF-998] adjust some comments and typos

* 👌 [RUMF-998] add a few tests to ensure the setting is correctly applied

* 👌 improve configuration value check

* More explicit function naming
* Improved typings inference

* 👌 use the InitialPrivacyLevel to declare the NodePrivacyLevel constants

* [RUMF-998] export the InitialPrivacyLevel constants

* 🚩 [RUMF-998] put the option behind a feature flag

* 👌 rename derivePrivacyLevelGivenParent to reducePrivacyLevel

* 👌 early return in `getNodePrivacyLevel`

Also, leverage the typegard and use `node` as an element directly

* 📝👌 adjust comments and doc