🔥 remove unused function `report()` and associated
The `report` function is used to report errors manually, similarly to
RUM `addError`. The SDK code doesn't use it, it is only used during
tests, but still included in the final bundle because the stacktrace
parsing logic is checking if a stack frame originated from the `report`
function.

Because this function uses a global state (to handle IE specificities,
which we don't do ourself), it is a bit problematic to make tracekit
code self contained and "stoppable".

This commit remove this function and associated code. In tests, we are
now actually throwing exceptions. Because jasmine fail tests if
exceptions are thrown during their execution, a utility helper has been
implemented to opt-out of this jasmine behavior.