🔊 [RUMF-1041] add initial visibilityState to lcp infos
This will help confirming the assumption that LCP value is not sent by
the browser when the tab is not visible. The 'discard_reason' is not
enough, because it is filled only when a LCP value is emitted (and
discarded), but in most cases it seems that the LCP value is not emitted
at all.