🐛 [RUMF-1077] use cookies to get Synthetics context
By using both global variable and cookies, the Synthetics context
retrieval should be more reliable. We shouldn't need to check it every
time we assemble an event. It should allow us to get it once at init,
and keep the value locally so we don't need caching mechanism to access
cookies and ensure good performances.