👷 remove as many disabled ESLint rules as possible
To simplify the ESLint configuration, remove as many explicitly disabled
rules that don't yield any error.

The `no-fallthourg` rule is the only removal that yield an error that
have been fixed. This is a good rule IMO, and the error was on the
tooling, so not much impact on production development.