♻️ [RUMF-1097] revamp configuration - rum (#1221)
* 🚚 [RUMF-1097][rum] move RumInitConfiguration and HybridConfiguration

* ♻️ [RUMF-1097][rum] expose a new RumConfiguration replacing Configuration

* ♻️ [RUMF-1097][rum] use the core validateAndBuildConfiguration

* ♻️ [RUMF-1097][rum] add applicationId into RumConfiguration

This allows to remove RumInitConfiguration usages outside of init() and
configuration building.

* 🔥 [RUMF-1097][core] delete previous buildConfiguration function

* 🏷️ fix TS compatibility

`ReadonlyArray<Foo>` is compiled by TS to `readonly Foo[]`, which is a
syntax introduced in TS 3.4[1]. Because TS compatibility will be updated
in v4 only, let's not use ReadonlyArray just yet. Having a writable Array here is
not a big deal anyway.

[1]: https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-4.html#improvements-for-readonlyarray-and-readonly-tuples

* 👌 simplify configuration build

This commit changes how we validate and build the configuration to make
it more concise. Since the default values are directly set during build,
we don't have a DEFAULT_CONFIGURATION object anymore. In tests, let's
use the function to build the configuration directly, so we get proper
Configuration objects.

* 👷 set `allow_failure: true` to check-squash-into-staging