👌 apply the rule on test files
This commit changes the rule a bit to make sure we only consider
directories actually containing an index file as private. Directories
containing a package.json file are not considered as private anymore, so
we can import from packages/core/test/... even if
packages/core/package.json exists.

It also adds the possibility to provide a glob pattern of index files to
ignore.

This allows us to apply the rule on test files.