🔥 [RUMF-1198]  remove problematic and useless code from the DeflateWorker (#1378)
* 🚨 [RUMF-1198] enable eslint on deflateWorker.js

I needed to reformat some comments to comply with the line size limit.
While those comments are cryptic and probably not helpful for us, they
are present in Pako source code. I chose to keep them, but maybe we
could remove them instead.

* [RUMF-1198] 🔥 remove some useless code from the DeflateWorker

This commit removes the ability to pass options to the `Deflate` class.
It's fine, because we never do. This allows to remove some problematic
"helper" functions causing issues when customers are transpiling the RUM
package with babel.