👷 [RUMF-1156] fix BrowserStack configuration
The previous browsers configuration was not
working because the new wdio packages are validating it (both with
typechecking and runtime checks).

With this commit, I try to improve how we define and pass browser
configuration to both karma and e2e tests. I also made sure that we are
running on the latest appium and selenium versions, to keep things up to
date.