✅ [RUMF-1156] use a bigger margin for network error tests
Apparently, with new wdio packages, the browser is reading more data
from request bodies at once (maybe because of a bigger buffer
somewhere), and the margin used to check that the body is not read
entirely was not enough. Still lower than the whole response, so the
test is still ensuring that the whole response is not read.