✨ [RUMF-1207] finally, collect concurrent actions!
Finally, remove the "one action at a time" limitation and collect a new
action for every click, even if another one is ongoing.

To do this, we still need to keep a reference to pending actions because
we need to discard them:
* when a view is created (this will be removed in a future PR)
* when stopping action collections (for tests)

This is somewhat cumbersome, but not a huge deal either.