♻️ [RUMF-1207] do some light refactoring in trackActions
This commit moves the "currentAction" ownership to the main
`trackActions` function. This will allow future commits to iterate over
the concept of "currentAction" and expose an API to access this
information.

To keep things more or less simple, the event listener logic has been
moved away from the `trackActions` function.

The `PendingAutoAction` class has been replaced with a `createAction`
function, to follow the trend of preferring function over classes. It is
also in charge of waiting the page to be idle, to make it
self-contained: the whole action lifecycle is now in the same function.
This will make it easier to track concurrent actions.

This design might change drastically in "frustration-signals" future
tasks.