👷 [RUMF-1156] update wdio to latest version  (#1405)
* 👷 [RUMF-1156] update wdio to latest version

* 🏷️ [RUMF-1156] work around stricter webdriver types

* ✅ [RUMF-1156] use a bigger margin for network error tests

Apparently, with new wdio packages, the browser is reading more data
from request bodies at once (maybe because of a bigger buffer
somewhere), and the margin used to check that the body is not read
entirely was not enough. Still lower than the whole response, so the
test is still ensuring that the whole response is not read.

* 👷 [RUMF-1156] fix BrowserStack configuration

The previous browsers configuration was not
working because the new wdio packages are validating it (both with
typechecking and runtime checks).

With this commit, I try to improve how we define and pass browser
configuration to both karma and e2e tests. I also made sure that we are
running on the latest appium and selenium versions, to keep things up to
date.

* 👌 add some doc and remove a test case

`capabilities.platformName` is never defined with our setup, so let's
remove it to simplify the function.