♻️ [RUMF-1207] collect concurrent actions groundwork - move action history closer to action collection (#1432)
* 🔥 remove unneeded parent context update

Previously, we updated the view parent context because it included the
current URL. Now that we have a dedicated `UrlContext` for this, we
don't need to update the view context anymore: all values are immutable
(view id and name).

* ♻️ [RUMF-1207] do some light refactoring in trackActions

This commit moves the "currentAction" ownership to the main
`trackActions` function. This will allow future commits to iterate over
the concept of "currentAction" and expose an API to access this
information.

To keep things more or less simple, the event listener logic has been
moved away from the `trackActions` function.

The `PendingAutoAction` class has been replaced with a `createAction`
function, to follow the trend of preferring function over classes. It is
also in charge of waiting the page to be idle, to make it
self-contained: the whole action lifecycle is now in the same function.
This will make it easier to track concurrent actions.

This design might change drastically in "frustration-signals" future
tasks.

* ♻️ [RUMF-1207] use history to track current action

Instead of using a plain variable, let's use the ContextHistory class
directly! In future commits will use this history instead of the one in
parentContexts.

* ♻️ [RUMF-1207] expose and use actionContexts instead of parentContexts

* 🔥 [RUMF-1207] remove parentContexts.findAction

* 🔥 [RUMF-1207] remove lifecycle events

* 👌 rename createAction to newAction

* 👌 rename parentContexts to viewContexts

* 👌 move actionContexts to trackActions

* 🔥 remove duplicated ActionContexts type definition

* 👌 shorter action contexts definition